Closes #87 

### What changed

- Add sample common item struct and aplly to filterIterator.
- Add benchmark result.

### Others

----

### Checklist

- [ ] **Issue is linked.**
    - It is not needed if simple edit such as corrections of words, etc.
- [ ] **Test is added.**
    - If fixed bug, add case on existing test or add new test.
    - If add new feature, add new test.
    - Add benchmark test, as needed.
- [ ] **Test coverage rate has no decreased.**
    - If rate is decreased, explain why in What changed.