@meian 2022-02-05
以下の方法で計測(benchディレクトリ内)

- 変更前の処理で以下を実行
    - `go test -bench BenchmarkFilter_Volumes -benchmem -benchtime 5s > without-struct.txt`
- 変更後の処理で以下を実行
    - `go test -bench BenchmarkFilter_Volumes -benchmem -benchtime 5s > with-struct.txt`
- 比較
    - `benchcmp without-struct.txt with-struct.txt` 

https://github.com/meian/gcf/blob/c3f8a86797215d8115f8f517a550d832ec549fba/bench/benchstat.txt

```
name                                    old time/op    new time/op    delta
Filter_Volumes/slice=10/filter=1-8         199ns ± 0%     199ns ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=5-8         701ns ± 0%     755ns ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=10-8       1.33µs ± 0%    1.38µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=100-8      11.3µs ± 0%    11.5µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=1-8       1.01µs ± 0%    1.02µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=5-8       4.57µs ± 0%    4.83µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=10-8      8.53µs ± 0%    8.70µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=100-8     70.8µs ± 0%    68.8µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=1-8      9.01µs ± 0%    9.20µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=5-8      44.9µs ± 0%    46.6µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=10-8     80.5µs ± 0%    80.6µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=100-8     667µs ± 0%     638µs ± 0%   ~     (p=1.000 n=1+1)

name                                    old alloc/op   new alloc/op   delta
Filter_Volumes/slice=10/filter=1-8         80.0B ± 0%     96.0B ± 0%   ~     (p=1.000 n=1+1) => +20.0%
Filter_Volumes/slice=10/filter=5-8          208B ± 0%      288B ± 0%   ~     (p=1.000 n=1+1) => +38.5%
Filter_Volumes/slice=10/filter=10-8         368B ± 0%      528B ± 0%   ~     (p=1.000 n=1+1) => +43.5%
Filter_Volumes/slice=10/filter=100-8      3.25kB ± 0%    4.85kB ± 0%   ~     (p=1.000 n=1+1) => +49.2%
Filter_Volumes/slice=100/filter=1-8        80.0B ± 0%     96.0B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=5-8         208B ± 0%      288B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=10-8        368B ± 0%      528B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=100-8     3.25kB ± 0%    4.85kB ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=1-8       80.0B ± 0%     96.0B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=5-8        208B ± 0%      288B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=10-8       368B ± 0%      528B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=100-8    3.25kB ± 0%    4.85kB ± 0%   ~     (p=1.000 n=1+1)

name                                    old allocs/op  new allocs/op  delta
Filter_Volumes/slice=10/filter=1-8          2.00 ± 0%      2.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=5-8          6.00 ± 0%      6.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=10-8         11.0 ± 0%      11.0 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=100-8         101 ± 0%       101 ± 0%   ~     (all equal)
Filter_Volumes/slice=100/filter=1-8         2.00 ± 0%      2.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=100/filter=5-8         6.00 ± 0%      6.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=100/filter=10-8        11.0 ± 0%      11.0 ± 0%   ~     (all equal)
Filter_Volumes/slice=100/filter=100-8        101 ± 0%       101 ± 0%   ~     (all equal)
Filter_Volumes/slice=1000/filter=1-8        2.00 ± 0%      2.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=1000/filter=5-8        6.00 ± 0%      6.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=1000/filter=10-8       11.0 ± 0%      11.0 ± 0%   ~     (all equal)
Filter_Volumes/slice=1000/filter=100-8       101 ± 0%       101 ± 0%   ~     (all equal)
```

処理速度はほぼ変わらないけど、噛ませる処理を増やすとメモリの割当が多くなる感じ？
気になるのでもう少し処理数増やしてみる