@meian 2022-02-05
アロケーションはスライスのサイズに依存しないのでslice=10のみで比較。

- Filter1000個を追加した比較結果。
    - https://github.com/meian/gcf/blob/1c85e056bf69d7a5f8f81dccf07bfba0e99d4b7c/bench/benchstat.txt

```
name                                   old time/op    new time/op    delta
Filter_Volumes/slice=10/filter=1-8        210ns ± 0%     209ns ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=5-8        739ns ± 0%     800ns ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=10-8      1.39µs ± 0%    1.45µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=100-8     11.8µs ± 0%    12.2µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=1000-8     123µs ± 0%     134µs ± 0%   ~     (p=1.000 n=1+1)

name                                   old alloc/op   new alloc/op   delta
Filter_Volumes/slice=10/filter=1-8        80.0B ± 0%     96.0B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=5-8         208B ± 0%      288B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=10-8        368B ± 0%      528B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=100-8     3.25kB ± 0%    4.85kB ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=1000-8    32.0kB ± 0%    48.0kB ± 0%   ~     (p=1.000 n=1+1)

name                                   old allocs/op  new allocs/op  delta
Filter_Volumes/slice=10/filter=1-8         2.00 ± 0%      2.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=5-8         6.00 ± 0%      6.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=10-8        11.0 ± 0%      11.0 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=100-8        101 ± 0%       101 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=1000-8     1.00k ± 0%     1.00k ± 0%   ~     (all equal)
```

- 共通項目の構造体なし
    - https://github.com/meian/gcf/blob/1c85e056bf69d7a5f8f81dccf07bfba0e99d4b7c/bench/without-struct.txt

```
goos: linux
goarch: amd64
pkg: github.com/meian/gcf/bench
cpu: Intel(R) Core(TM) i5-8250U CPU @ 1.60GHz
BenchmarkFilter_Volumes/slice=10/filter=1-8         	24976196	       210.2 ns/op	      80 B/op	       2 allocs/op
BenchmarkFilter_Volumes/slice=10/filter=5-8         	 7946065	       738.8 ns/op	     208 B/op	       6 allocs/op
BenchmarkFilter_Volumes/slice=10/filter=10-8        	 4252044	      1394 ns/op	     368 B/op	      11 allocs/op
BenchmarkFilter_Volumes/slice=10/filter=100-8       	  446185	     11833 ns/op	    3248 B/op	     101 allocs/op
BenchmarkFilter_Volumes/slice=10/filter=1000-8      	   48360	    122594 ns/op	   32048 B/op	    1001 allocs/op
PASS
ok  	github.com/meian/gcf/bench	36.968s
```

- 共通項目の構造体あり
    - https://github.com/meian/gcf/blob/1c85e056bf69d7a5f8f81dccf07bfba0e99d4b7c/bench/with-struct.txt

```
goos: linux
goarch: amd64
pkg: github.com/meian/gcf/bench
cpu: Intel(R) Core(TM) i5-8250U CPU @ 1.60GHz
BenchmarkFilter_Volumes/slice=10/filter=1-8         	26863389	       209.4 ns/op	      96 B/op	       2 allocs/op
BenchmarkFilter_Volumes/slice=10/filter=5-8         	 7438172	       800.5 ns/op	     288 B/op	       6 allocs/op
BenchmarkFilter_Volumes/slice=10/filter=10-8        	 4100841	      1448 ns/op	     528 B/op	      11 allocs/op
BenchmarkFilter_Volumes/slice=10/filter=100-8       	  454443	     12200 ns/op	    4848 B/op	     101 allocs/op
BenchmarkFilter_Volumes/slice=10/filter=1000-8      	   45298	    133687 ns/op	   48048 B/op	    1001 allocs/op
PASS
ok  	github.com/meian/gcf/bench	33.146s
```

個別の結果を見て気づいたけど、

- 既存の実装はFilter毎に32バイトのアロケーションが発生
- structをつけるとFilter毎に48バイトのアロケーションが発生

となっていた。