@meian 2022-02-05
```golang
type iteratorItem[T any] struct {
	current T
	done    bool
}

↓↓↓↓↓

type iteratorItem[T any] struct {
	current T
}
```

こう変えたらまた32バイト増加に戻った。  
bool一つ増やして16バイト増えるのはなぜ？  
この辺はよくわからないけどもともとdoneフラグある処理もあるので、基本的に何も増えることはなさそう